# swagger.yaml
openapi: 3.0.3
info:
  title: Kitty Party App API
  description: |
    REST API for managing Indian-style kitty party savings pools.

    Members contribute a fixed amount each month; one member hosts and
    receives the pooled amount each cycle. This API covers the full
    lifecycle: members, groups & memberships, profiles, kitty cycles,
    monthly dues, payments, and payouts.

    All successful responses use the envelope `{ "success": true, "data": ... }`.
    All errors use `{ "success": false, "error": { "code": <http>, "message": "..." } }`.
  version: "1.0.0"
  contact:
    name: Kitty Party Dev
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /api/v1/members:
    get:
      tags: [Members]
      summary: List all members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListEnvelope"
    post:
      tags: [Members]
      summary: Create a member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMemberRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/members/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Members]
      summary: Get member by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Members]
      summary: Update a member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMemberRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Members]
      summary: Delete a member
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Profiles ───────────────────────────────────────────────────────────────
...
  /api/v1/groups:
    get:
      tags: [Groups]
      summary: List all groups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupListEnvelope" }
    post:
      tags: [Groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGroupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/groups/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Groups]
      summary: Get a group
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Groups]
      summary: Update a group
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateGroupRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Groups]
      summary: Delete a group
      responses:
        "204": { description: Deleted }
...
  /api/v1/groups/{id}/members:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Memberships]
      summary: List members in a group
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupMembershipListEnvelope" }
    post:
      tags: [Memberships]
      summary: Add a member to a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddMemberRequest" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupMembershipEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/groups/{id}/members/count:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Memberships]
      summary: Count members in a group
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count: { type: integer, example: 7 }

  # ─── Contributions ──────────────────────────────────────────────────────────
  /api/v1/contributions/dues/generate:
    post:
      tags: [Contributions]
      summary: Fan out monthly dues for every ACTIVE group member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateDuesRequest" }
      responses:
        "201":
          description: Dues generated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionDueListEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/contributions/groups/{groupID}/dues:
    parameters:
      - in: path
        name: groupID
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: cycle_month
        schema: { type: integer, minimum: 1, maximum: 12 }
    get:
      tags: [Contributions]
      summary: List dues for a group (optionally filtered by cycle_month)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionDueListEnvelope" }
  /api/v1/contributions/payments:
    post:
      tags: [Contributions]
      summary: Record a payment against a due
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecordPaymentRequest" }
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionPaymentEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ─── Contribution Tracking ──────────────────────────────────────────────────
  /api/v1/contribution-tracking:
    post:
      tags: [Contribution Tracking]
      summary: Initialize a per-member, per-cycle tracking row
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTrackingRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/contribution-tracking/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Contribution Tracking]
      summary: Get a tracking row by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/contribution-tracking/{id}/payment:
    parameters:
      - $ref: "#/components/parameters/PathID"
    patch:
      tags: [Contribution Tracking]
      summary: Apply a payment delta to a tracking row
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTrackingRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/contribution-tracking/dues/{dueID}:
    parameters:
      - in: path
        name: dueID
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Contribution Tracking]
      summary: Get the tracking row tied to a specific due
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/contribution-tracking/groups/{groupID}:
    parameters:
      - in: path
        name: groupID
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: cycle_month
        schema: { type: integer, minimum: 1, maximum: 12 }
      - in: query
        name: cycle_year
        schema: { type: integer, minimum: 2000 }
    get:
      tags: [Contribution Tracking]
      summary: List tracking rows for a group (optionally filtered by cycle)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingListEnvelope" }
  /api/v1/contribution-tracking/members/{memberID}:
    parameters:
      - in: path
        name: memberID
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Contribution Tracking]
      summary: List tracking rows for a member across groups/cycles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContributionTrackingListEnvelope" }

  # ─── Payouts ────────────────────────────────────────────────────────────────
...
  /api/v1/payouts/groups/{groupID}:
    parameters:
      - in: path
        name: groupID
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Payouts]
      summary: List payouts for a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayoutSummaryListEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/kitty-cycles/groups/{groupID}:
    parameters:
      - in: path
        name: groupID
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Kitty Cycles]
      summary: List cycles for a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KittyCycleListEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Issue a token via `POST /api/v1/auth/token`.
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation failed or malformed payload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
    Forbidden:
      description: Caller is not permitted to perform this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
    Conflict:
      description: Duplicate or invalid-state operation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
...
  schemas:
...
    APIResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
        data: {}
    APIErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: integer, example: 400 }
            message: { type: string, example: "validation failed: ..." }

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Member:
      type: object
      properties:
        id:         { type: string, format: uuid }
        name:       { type: string }
        email:      { type: string, format: email }
        phone:      { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateMemberRequest:
      type: object
      required: [name, email, phone]
      properties:
        name:  { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 10, maxLength: 15 }
    UpdateMemberRequest:
      type: object
      properties:
        name:  { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 10, maxLength: 15 }
    MemberEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Member" }
    MemberListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Member" }

    # ── Profile ─────────────────────────────────────────────────────────────
...
    Group:
      type: object
      properties:
        id:             { type: string, format: uuid }
        name:           { type: string }
        monthly_amount: { type: number, format: double }
        duration:       { type: integer, description: "Number of months" }
        start_date:     { type: string, format: date-time }
        created_by:     { type: string, format: uuid }
        created_at:     { type: string, format: date-time }
        updated_at:     { type: string, format: date-time }
    CreateGroupRequest:
      type: object
      required: [name, monthly_amount, duration, start_date, created_by]
      properties:
        name:           { type: string, minLength: 2, maxLength: 100 }
        monthly_amount: { type: number, format: double, exclusiveMinimum: 0 }
        duration:       { type: integer, minimum: 1 }
        start_date:     { type: string, format: date-time }
        created_by:     { type: string, format: uuid }
    UpdateGroupRequest:
      type: object
      properties:
        name:           { type: string, minLength: 2, maxLength: 100 }
        monthly_amount: { type: number, format: double, exclusiveMinimum: 0 }
        duration:       { type: integer, minimum: 1 }
        start_date:     { type: string, format: date-time }
    GroupMembership:
      type: object
      properties:
        id:          { type: string, format: uuid }
        group_id:    { type: string, format: uuid }
        member_id:   { type: string, format: uuid }
        member_name: { type: string }
        status:      { type: string, enum: [ACTIVE, PENDING, REJECTED] }
        role:        { type: string, enum: [MEMBER, ADMIN] }
        joined_at:   { type: string, format: date-time }
    AddMemberRequest:
      type: object
      required: [member_id]
      properties:
        member_id: { type: string, format: uuid }
...
    GroupEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Group" }
    GroupListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Group" }
    GroupMembershipEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/GroupMembership" }
    GroupMembershipListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/GroupMembership" }
...
    ContributionDue:
      type: object
      properties:
        id:               { type: string, format: uuid }
        group_id:         { type: string, format: uuid }
        member_id:        { type: string, format: uuid }
        host_member_id:   { type: string, format: uuid }
        cycle_month:      { type: integer, minimum: 1, maximum: 12 }
        cycle_year:       { type: integer, minimum: 2000 }
        cycle_number:     { type: integer, minimum: 1 }
        kitty_party_date: { type: string, format: date-time }
        due_date:         { type: string, format: date-time }
        amount:           { type: number, format: double }
        status:           { type: string, enum: [PENDING, PAID, OVERDUE] }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time, nullable: true }
    ContributionPayment:
      type: object
      properties:
        id:              { type: string, format: uuid }
        due_id:          { type: string, format: uuid }
        amount_paid:     { type: number, format: double }
        payment_date:    { type: string, format: date-time }
        payment_mode:    { type: string, enum: [UPI, CASH, BANK] }
        transaction_ref: { type: string, nullable: true }
        status:          { type: string, enum: [SUCCESS, FAILED] }
        created_at:      { type: string, format: date-time }
    GenerateDuesRequest:
      type: object
      required:
        - group_id
        - host_member_id
        - cycle_month
        - cycle_year
        - cycle_number
        - kitty_party_date
        - due_date
        - amount
      properties:
        group_id:         { type: string, format: uuid }
        host_member_id:   { type: string, format: uuid }
        cycle_month:      { type: integer, minimum: 1, maximum: 12 }
        cycle_year:       { type: integer, minimum: 2000 }
        cycle_number:     { type: integer, minimum: 1 }
        kitty_party_date: { type: string, format: date-time }
        due_date:         { type: string, format: date-time }
        amount:           { type: number, format: double, exclusiveMinimum: 0 }
    RecordPaymentRequest:
      type: object
      required: [due_id, amount_paid, payment_mode, status]
      properties:
        due_id:          { type: string, format: uuid }
        amount_paid:     { type: number, format: double, exclusiveMinimum: 0 }
        payment_mode:    { type: string, enum: [UPI, CASH, BANK] }
        transaction_ref: { type: string, nullable: true }
        status:          { type: string, enum: [SUCCESS, FAILED] }
    ContributionDueListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/ContributionDue" }
    ContributionPaymentEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ContributionPayment" }

    # ── Contribution Tracking ───────────────────────────────────────────────
    ContributionTracking:
      type: object
      properties:
        id:                 { type: string, format: uuid }
        group_id:           { type: string, format: uuid }
        member_id:          { type: string, format: uuid }
        due_id:             { type: string, format: uuid }
        cycle_month:        { type: integer, minimum: 1, maximum: 12 }
        cycle_year:         { type: integer, minimum: 2000 }
        cycle_number:       { type: integer, minimum: 1 }
        total_due:          { type: number, format: double }
        total_paid:         { type: number, format: double }
        balance:            { type: number, format: double }
        status:             { type: string, enum: [PENDING, PARTIAL, PAID, OVERDUE] }
        last_payment_date:  { type: string, format: date-time, nullable: true }
        created_at:         { type: string, format: date-time }
        updated_at:         { type: string, format: date-time }
    CreateTrackingRequest:
      type: object
      required:
        - group_id
        - member_id
        - due_id
        - cycle_month
        - cycle_year
        - cycle_number
        - total_due
      properties:
        group_id:     { type: string, format: uuid }
        member_id:    { type: string, format: uuid }
        due_id:       { type: string, format: uuid }
        cycle_month:  { type: integer, minimum: 1, maximum: 12 }
        cycle_year:   { type: integer, minimum: 2000 }
        cycle_number: { type: integer, minimum: 1 }
        total_due:    { type: number, format: double, exclusiveMinimum: 0 }
    UpdateTrackingRequest:
      type: object
      required: [amount_paid]
      properties:
        amount_paid:       { type: number, format: double, exclusiveMinimum: 0 }
        last_payment_date: { type: string, format: date-time, nullable: true }
    ContributionTrackingEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ContributionTracking" }
    ContributionTrackingListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/ContributionTracking" }

    # ── Payouts ─────────────────────────────────────────────────────────────
...
    PayoutSummary:
      allOf:
        - $ref: "#/components/schemas/PayoutSchedule"
        - type: object
          properties:
            disbursement:
              $ref: "#/components/schemas/Disbursement"
...
    PayoutSummaryListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/PayoutSummary" }

    # ── Kitty Cycles ────────────────────────────────────────────────────────
...
    KittyCycleListEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/KittyCycle" }
...